refactor(modules): seperating request api to a single module
