+++ send_module/__init__.py
@@ -1,0 +1,5 @@
+from . import util
+
+__all__ = [
+    "util"
+]
